# 176: (179, 214, 181, 182)
Set up the back end to verify if the user's email exists in the database. If it does, then they can begin to confirm their phone number and reset their password.

Additionally, the we set up the backend to find the user's email using their personal info and display that email if it's found.